--- admin/tpls/index.css
@@ -1,141 +1,0 @@
-<style type="text/css">
-#qtoppage {
-height: 650px;
-}
-#indexmenu a {
-	display: block;
-	float: left;
-	height: 95px !important;
-	height: 95px;
-	width: 90px !important;
-	width: 90px;
-	vertical-align: middle;
-	text-decoration: none;
-	background-color: #f0f0f0;
-	padding: 2px 5px 1px 5px;
-	margin: 3px;
-	color: #666666;
-	border: 1px solid #f9f9f9;
-	-moz-border-radius: 6px;
-	-webkit-border-radius: 6px;
-	-khtml-border-radius: 6px;
-	border-radius: 6px;
-}
-
-#indexmenu a:hover {
-	color: #1E90FF;
-	border-left: 1px solid #eee;
-	border-top: 1px solid #eee;
-	border-right: 1px solid #ccc;
-	border-bottom: 1px solid #ccc;
-	background: #f9f9f9;
-}
-#aca {height: 300px !important;}
-/*==============*/
-.menu_list_q {
-	width: 100%;
-}
-
-.menu_head_q {
-	padding: 5px 10px;
-	cursor: pointer;
-	position: relative;
-	margin: 1px;
-	background: #000 !important;
-	color: #FFFFFF;
-	border: 1px solid #FFF;
-	cursor: pointer;
-	font-weight: bold;
-}
-
-.menu_body_q {
-	display: none;
-	background: #f1f1f1;
-}
-
-.menu_body_q a {
-	display: block;
-	color: #3E728F;
-	padding-left: 10px;
-	font-weight: bold;
-	text-decoration: none;
-}
-
-.menu_body_q a:hover {
-	color: #000000;
-	text-decoration: none;
-}
-/*================================*/
-
-.tabbed_content {
-	background-color: #f7f7f7;
-	width: 100%;	
-}
-
-.tabs {
-	height: 62px;
-	position: relative;
-	background-color: #000;
-}
-
-.tabs .moving_bg {
-	padding-top: 15px;
-	padding-bottom: 15px;
-	background-color:#7F822A;
-	background-image:url(../assets/images/arrow_down_green.png);
-	position: absolute;
-	width: 90px;
-	z-index: 190;
-	left: 0;
-	padding-bottom: 29px;
-	background-position: bottom left;
-	background-repeat: no-repeat;
-}
-
-.tabs .tab_item {
-	display: block;
-	float: left;
-	padding-top: 15px;
-	padding-bottom: 15px;
-	width: 90px;
-	color: #ffffff;
-	text-align: center;
-	z-index: 200;
-	position: relative;
-	cursor: pointer;
-}
-
-.tabbed_content .slide_content {
-	overflow: hidden;
-	background-color: #f7f7f7;
-	padding: 20px 0 20px 20px;
-	position: relative;
-	width: 600px;
-}
-
-.tabslider {
-	max-width: 5000px;
-	height: 400px;
-}
-
-.tabslider ul {
-	float: left;
-	width: 560px;
-	margin: 0px;
-	padding: 0px;
-	margin-right: 40px;
-}
-
-.tabslider ul a {
-	color: #000;
-	text-decoration: none;
-}
-
-.tabslider ul a:hover {
-	color: #aaaaaa;
-}
-
-.tabslider ul li {
-	padding-bottom: 7px;
-}
-</style>
