+++ assets/images/index.html
@@ -1,1 +1,1 @@
-<script type="text/javascript">history.go(-1);</script>
+<script>history.go(-1);</script>
