+++ include/install.php
@@ -16,23 +16,21 @@
  * @author 			Mojtaba Jamali(jamali.mojtaba@gmail.com)
  * @version         $Id: $
  */
-function xquiz_CleanVars($global, $key, $default = '', $type = 'int')
+function xoops_module_install_xquiz()
 {
-    switch ($type) {
-        case 'string':
-            //$ret = (isset($global[$key])) ? $global[$key] : $default;
-            $ret = isset($global[$key]) ? filter_var($global[$key], FILTER_SANITIZE_ADD_SLASHES) : $default;
-            break;
-        case 'int':
-        default:
-            //$ret = (isset($global[$key])) ? intval($global[$key]) : intval($default);
-            $ret = isset($global[$key]) ? filter_var($global[$key], FILTER_SANITIZE_NUMBER_INT) : $default;
-            break;
+    $dir = XOOPS_ROOT_PATH . '/uploads/xquiz';
+
+    if (!is_dir($dir)) {
+        mkdir($dir, 0777);
+
+        chmod($dir, 0777);
     }
 
-    if (false === $ret) {
-        return $default;
-    }
+    $dir = XOOPS_ROOT_PATH . '/uploads/xquiz/image';
 
-    return $ret;
+    if (!is_dir($dir)) {
+        mkdir($dir, 0777);
+
+        chmod($dir, 0777);
+    }
 }
