+++ language/english/blocks.php
@@ -1,7 +1,38 @@
 <?php
+/**
+ * xquiz - MODULE FOR XOOPS
+ * Copyright (c) Mojtaba Jamali of persian xoops project (http://www.irxoops.org/)
+ *
+ * You may not change or alter any portion of this comment or credits
+ * of supporting developers from this source code or any supporting source code
+ * which is considered copyrighted (c) material of the original comment or credit authors.
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ *
+ * @copyright       XOOPS Project (https://xoops.org)
+ * @license         GNU GPL 2 (http://www.gnu.org/licenses/old-licenses/gpl-2.0.html)
+ * @package         xquiz
+ * @author 			Mojtaba Jamali(jamali.mojtaba@gmail.com)
+ * @version         $Id: $
+ */
+define('_MB_XQUIZ_OP1', '<b>Quiz width</b>');
+define('_MB_XQUIZ_OP2', '<b>Quiz height</b>');
+define('_MB_XQUIZ_OP3', '<b>Image width</b>');
+define('_MB_XQUIZ_OP4', '<b>Image height</b>');
+define('_MB_XQUIZ_OP5', '<b>Select Quiz category</b>');
+define('_MB_XQUIZ_OP6', '<b>Select Quiz type</b>');
+define('_MB_XQUIZ_OP7', '<b>Limit</b>');
+define('_MB_XQUIZ_OP8', '<b>Display Caption?</b>');
+define('_MB_XQUIZ_OP9', '<b>Slider Style Class</b>');
+define('_MB_XQUIZ_NIVOSLIDER', 'Nivo Quiz');
+define('_MB_XQUIZ_SLICKSLIDER', 'Slick Quiz');
+define('_MB_XQUIZ_STYLE', '<b>Use module style, if it set to no you must add all styles and java in your theme</b>');
+
 define('_MB_XQUIZ_ACTIVE', 'Active');
 define('_MB_XQUIZ_UNACTIVE', 'Inactive');
 define('_MB_XQUIZ_QUESTION', 'Question');
 define('_MB_XQUIZ_OPTION', 'Total of Quiz to be Display');
 define('_MB_XQUIZ_LATESTQUIZ', 'Latest Quiz');
 define('_MB_XQUIZ_LATESTQUIZ_ACTIVE', 'Latest Active Quiz');
+
