+++ language/english/main.php
@@ -1,4 +1,23 @@
 <?php
+/**
+ * xquiz - MODULE FOR XOOPS
+ * Copyright (c) Mojtaba Jamali of persian xoops project (http://www.irxoops.org/)
+ *
+ * You may not change or alter any portion of this comment or credits
+ * of supporting developers from this source code or any supporting source code
+ * which is considered copyrighted (c) material of the original comment or credit authors.
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ *
+ * @copyright       XOOPS Project (https://xoops.org)
+ * @license         GNU GPL 2 (http://www.gnu.org/licenses/old-licenses/gpl-2.0.html)
+ * @package         xquiz
+ * @author 			Mojtaba Jamali(jamali.mojtaba@gmail.com)
+ * @version         $Id: $
+ */
+define('_MD_XQUIZ_DEFAULT', 'xQuiz');
+
 define("_MD_XQUIZ_MODULENAME", "Quiz");
 define("_AC_COPYRIGHT", "M-Jamali");
 define('_QUEST_NUMBER', 'Questions number');
@@ -64,3 +83,4 @@ define("_XQUIZ_EMAIL_DESC", "Dear %s, this is a message transfered from your acc
 define('_XQUIZ_EMAIL_MESSAGE', 'Email message');
 define('_XQUIZ_SCORE_PROFILE', 'More details in your quiz profile');
 define('_XQUIZ_EMAIL_SUBJECT', 'Quiz result');
+
