+++ templates/admin/xquiz_header.tpl
@@ -1,0 +1,9 @@
+<div class="moduleicons">
+    <div class="floatright">
+        <div class="xo-buttons">
+            <a class="ui-corner-all tooltip" href="quiz.php?op=new_item" title="<{$smarty.const._AM_XQUIZ_ADD_XQUIZ}>"><img src="<{xoAdminIcons add.png}>" alt="<{$smarty.const._AM_XQUIZ_ADD_XQUIZ}>"/><{$smarty.const._AM_XQUIZ_ADD_XQUIZ}></a>
+            <a class="ui-corner-all tooltip" href="category.php?op=new_category" title="<{$smarty.const._AM_XQUIZ_ADD_CATEGORY}>"><img src="<{xoAdminIcons folder_blue.png}>" alt="<{$smarty.const._AM_XQUIZ_ADD_CATEGORY}>"/><{$smarty.const._AM_XQUIZ_ADD_CATEGORY}></a>
+        </div>
+    </div>
+    <div class="navigation" id="<{$navigation}>"><{$navtitle}></div>
+</div>
