+++ templates/admin/xquiz_index.tpl
@@ -1,0 +1,2 @@
+<{includeq file="$xoops_rootpath/modules/xquiz/templates/admin/xquiz_header.tpl"}>
+<{$renderindex}>
